revert "upgrade", use "remote --outdated" instead
for some reason, "vcpkg upgrade" removes the port's folder in the
packages/ directory, but cesium's cmake rely on it to install headers
and libs!
This is a bad idea because when cleaning the (ez)vcpkg folder and
letting vcpkg re-install using "binarysource" (archives cache), only the
"installed" folder is populated, not the "packages", thus breaking the
build.
The header and library file list to install should rather be extracted
from the installed/vcpkg/info/<port>_<version>_<triplet>.list files
instead. Why the manual install step btw?